+++ centrifuge-app/src/components/Portfolio/CoinsSvg.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-export function Coins() {
+export function CoinsSvg() {
   return (
     <svg viewBox="0 0 119 72" fill="none" xmlns="http://www.w3.org/2000/svg">
       <path
