+++ centrifuge-app/src/components/Portfolio/InvestedTokens.tsx
@@ -1,8 +1,12 @@
-import { useBalances } from '@centrifuge/centrifuge-react'
+import { AccountTokenBalance, Pool } from '@centrifuge/centrifuge-js'
+import { formatBalance, useBalances } from '@centrifuge/centrifuge-react'
 import { Box, Grid, Stack, Text } from '@centrifuge/fabric'
 import * as React from 'react'
 import { useAddress } from '../../utils/useAddress'
-import { TokenCard, TOKEN_CARD_COLUMNS, TOKEN_CARD_GAP } from '../TokenCard'
+import { usePool } from '../../utils/usePools'
+
+const TOKEN_ITEM_COLUMNS = `250px 200px 100px 150px 1FR`
+const TOKEN_ITEM_GAP = 4
 
 export function InvestedTokens() {
   const address = useAddress()
@@ -16,7 +20,7 @@ export function InvestedTokens() {
         </Text>
       </Box>
       <Stack gap={1}>
-        <Grid gridTemplateColumns={TOKEN_CARD_COLUMNS} gap={TOKEN_CARD_GAP} px={2}>
+        <Grid gridTemplateColumns={TOKEN_ITEM_COLUMNS} gap={TOKEN_ITEM_GAP} px={2}>
           <Text as="span" variant="body3">
             Token
           </Text>
@@ -34,11 +38,52 @@ export function InvestedTokens() {
         <Stack as="ul" role="list" gap={1}>
           {balances.tranches.map((tranche, index) => (
             <Box key={`${tranche.trancheId}${index}`} as="li">
-              <TokenCard {...tranche} />
+              <TokenListItem {...tranche} />
             </Box>
           ))}
         </Stack>
       </Stack>
     </>
   ) : null
+}
+
+type TokenCardProps = AccountTokenBalance
+export function TokenListItem({ balance, currency, poolId, trancheId }: TokenCardProps) {
+  const pool = usePool(poolId) as Pool
+  const isTinlakePool = poolId?.startsWith('0x')
+
+  if (isTinlakePool) {
+    return null
+  }
+
+  const tranche = pool.tranches.find(({ id }) => id === trancheId)
+
+  return (
+    <Grid
+      gridTemplateColumns={TOKEN_ITEM_COLUMNS}
+      gap={TOKEN_ITEM_GAP}
+      padding={2}
+      borderStyle="solid"
+      borderWidth={1}
+      borderColor="borderSecondary"
+    >
+      <Text as="span" variant="body2">
+        {currency.name}
+      </Text>
+
+      <Text as="span" variant="body2">
+        {formatBalance(balance, tranche?.currency.symbol)}
+      </Text>
+
+      <Text as="span" variant="body2">
+        {tranche?.tokenPrice ? formatBalance(tranche.tokenPrice.toDecimal(), tranche.currency.symbol, 4) : '-'}
+      </Text>
+
+      <Text as="span" variant="body2">
+        {tranche?.tokenPrice
+          ? formatBalance(balance.toDecimal().mul(tranche.tokenPrice.toDecimal()), tranche.currency.symbol, 4)
+          : '-'}
+      </Text>
+    </Grid>
+  )
 }
